Enable to benchmark on CI
This commit enables us to measure container's startup performance with benchmark
based on HelloBench project. On each opening and updating PR, we run the
benchmark on the GithubAction's "ubuntu-18.04" runner and post the result as a
comment. We use Docker Hub for storing legacy/stargz/extended-stargz images.

See also "HelloBench" project:
- https://github.com/Tintri/hello-bench

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>