Resolve layers in parallel
Currently, resolving each layer in an image are done sequentially. But it can be
done in parallel by passing all layer digests in the image to `Prepare` of this
snapshotter.

This commit enables this by adding a new optional label
"containerd.io/snapshot/remote/stargz.chain" for passing all digests contained
in the target image.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>