Move config files for Dockerfile to separated directory
Currently config files used in the Dockerfile is contained in
`/script/pullsecrets/config` but these files are now used more commonly and used
also in the CRI test and in the quickstart in README. So these files should be
moved to more general location. This commit moves these config files to
`/script/config`.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>