Define log message for failure of remote snapshot preparation
Currently, we have two ways to check the success of remote snapshot
preparation. Integration test checks output messages from `ctr-remote`
command. CRI test checks the log from stargz daemon. We should unify them
because we want to keep the consideration the about log messages as small as
possible.

We use the log from stargz daemon for both tests. This commit formats all log
with JSON for ease of querying information and defines a special key for
indicating whether remote snapshot preparation suceeded or not.

This commit also adds this check to the benchmarking script.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>