Use single Dockerfile for version management of containerd in tests
Currently, each test has each Dockerfile and manages the version of containerd
and runc, etc. by itself, which is a bit costly for us in terms of version
management.

This commit refactors this by defining a common Dockerfile which defines the
versions of containerd and runc used in this project and which can be used as a
base image for testing containers. This commit also introduces "CinD"(Containerd
In Docker) build stage (image) in that Dockerfile which can be used for easy
demo or manual testing.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>