Add `Unmount` API to the FileSystem
Our `FileSystem` implementation holds a `map[string]*layer` for tracking the
state of each stargz mountpoint. Currently, `FileSystem` interface defined in
`snapshot` package doesn't have `Unmount` API and `syscall.Unmount` is directly
called to the mountpoint. Filesystem implementation asynchronously finalizes
that `map`, which is hard to be done safely, rather race condition will possibly
occur.

This is good timing for us to add `Unmount` method to the `FileSystem` API,
which delegates the actual `unmount` task to the filesystem implementation. This
also enables the filesystem to do necessary finalization before/after the
`unmount` syscall.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>