Extend stargz for content verification
Current stargz image has two problems towards content validation:
- The digest of TOC JSON isn't stored in verifiable image manifests
- Content digests are stored in file granularity (not in chunk granularity)

By solving them, we can construct a chain of verification (i.e. manifest -> TOC
JSON -> chunk).

This commit solves them in backwards-compatible ways as the following.

For the first one, this commit adds an annotation
`containerd.io/snapshot/stargz/toc.digest` to the image manifest for storing the
digest of TOC JSON.

For the second one, this commit extends the TOCEntry with an additional (but
optional) field `ChunkDigest` for storing chunk-level content digest. Stargz lib
chunks a large file to smaller "chunk" entries (4MiB by default). In stargz
snapshotter, we leverage this and fetch a large file in chunk granularity,
selectively. However, stargz lib __doesn't__ calculate the digest for each
chunk. Instead, it calculates the digest for the entire file and stores this
to`TOCEntry.Digest` of the first chunk. `TOCEntry.Digest` of other chunks are
kept empty. For keeping backwards-compatibility with stargz, this commit adds a
new field for containing chunk-level digests instead of changing the semantics
of `TOCEntry.Digest`.

Signed-off-by: Kohei Tokuanga <ktokunaga.mail@gmail.com>