Enable filesystem to verify stargz blob contents
Currently, filesystem doesn't verify the layer contents.
Using `verify` package, this commit enables filesystem to verify:

- TOC JSON of the stargz layer, and
- each content contained in that layer

The expected digest of the TOC JSON must be passed via label
("containerd.io/snapshot/stargz/toc.digest"). Currently, `ctr-remote` command
adds it to eStargz image as annotations and the annotation is passed down to
this snaphsotter.

For legacy stargz images that don't contain the TOC JSON digest annotations,
filesystem allows to mount them if either of the following conditions meets,

- `allow_no_verification` is set to `true` in the config.toml of this
  snapshotter, or
- label "containerd.io/snapshot/remote/stargz.skipverify" is appended to that
  layer.

This commit adds a flag to `ctr-remote image rpull` command to allow the
filesystem to deal with legacy stargz images, using the second method in the
above.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>