Show benchmarking result in more detail and improve the stability
In our current benchmark, we simply show the average time to take for each
operation during container startup.
This commit improves the benchmarking codes to contain the benchmarking results
in more detail. This commit also includes some changes trying to improve the
stability of the benchmarking results. We don't change the benchmarking method
itself and is still HelloBench-based one.

- Using 95 percentiles instead of the average value.
- Containing lists of percentile values of the benchmark results.
- Not executing `sync` command explicitly during benchmarking of `pull`
  operation. The time to take for (implicit) disk sync will be included in
  operations (create, run) following the pull operation.
- For each iteration during the benchmark, randomize the order of samples
  execution.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>