Make keychain selection configurable
Currently, kubeconfig-based keychain is imported to the filesystem by default
but this comes with a lot of dependencies (including `k8s.io` pkg) and this
makes importing `github.com/containerd/stargz-snapshotter/stargz` harder. Since
this commit, the filesystem doesn't import the keychain by default and the
keychain selection is configurable through the option. There is no change in
`config.toml`.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>