Docker CLI plugin for image optimization
We already have a containerized version of `ctr-remote images optimize`. This
commit enables to run this image from a docker CLI plugin.

This plugin runs `ctr-remote` command inside a privileged container and uses the
following host resources.
- `/dev/fuse` mapped from host
- host's .docker/config.json for registry authn

In the future, we should consider to make the plugin run totally in a rootless
manner.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>