containerd's RegistryHost-based registry logic
Currently, registry logic of stargz snapshotter is separated from containerd and
is implemented in a imcompatible way with other containerd-based tools
(including buildkit which uses containerd's docker.RegistryHost). This makes it
hard to integrate the logic with these tools.

This commit refactors registry logic based on containerd's
docker.RegistryHost. This enables the client of stargz.FileSystem to configure
docker.RegistryHost based on the client-side configuration. This will make the
integration of the registry logic with other containerd-based tools easier.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>