Allow client to configure blob URLs through RegistryHost configuration
This commit adds a new field `RefAlias` to the registry configuration. This
enables the client to configure the destination URL of layers more flexibly. For
example, the client can change the URL of a layer blob on refresh, based on the
connectivity.

This implies that a snapshot (or layer) shouldn't be tied to a single image
reference and we might need more refactoring on the registry logic but let's
keep this separated from this patch.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>