Enable ctr-remote to handle image index
Currently, ctr-remote only converts images matching to the platform it running
on.
This commit allows ctr-remote to handle image index and convert images of all or
a specific platform. Based on the ability of handling index, this commit also
enables to input/output imgaes from/to the local filesystem following OCI Image
Layout.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>