Replace the multierror dependency with a simple error aggregator.
This lets functions that want to aggregate and return a collection of errors use a `[]error` to accumulate them, and `errorutil.Aggregate(list)` to turn them into a single `error`.  If the list is empty, the `error` will be nil.  If the list is a singleton, it will return the single error.  Otherwise, it will construct a new error combining the constituent error messages.

Fixes: https://github.com/containerd/stargz-snapshotter/issues/205
Signed-off-by: Matt Moore <mattmoor@vmware.com>