Support absolute and root relative prioritized file paths
Since this commit, WithPrioritiedFiles option accepts absolute paths and paths
that rellative to root (e.g. paths prefixed with "./").
The input tar can include files that are relative to root (e.g. prefixed with
"./"). So, in this commit, the specified prioritized file paths are normalized
(using filepath.Clean()) and are compared against normalized tar entry names.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>