estargz: Clean entry names when read and unmarshal blob
Currently, when estargz lib reads an eStargz, it normalizes the entry by
trimming the prefix "./".

But this isn't enough because there are a lot of exceptions including
"/foo/bar", "../foo/bar", etc. From the library client's perspective
(e.g. snapshotter), it's complex to take care of all these differences when
interacting with unmarshaled eStargz blob with `estargz.Reader`.

For example, when the client gets a TOCEntry from `estargz.Reader.Lookup(path
string)`, it needs to prefix the path to be lexically equal to the corresponding
tar entry (so it needs to prefix that path with "/" if the corresponding tar
entry is prefixed by "/"). Additionally, when the client gets a TOCEntry
(e.g. by `estargz.TOCEntry.ForeachChild()`), the names are possibly prefixed
with "/", "./", "../", "", etc and it needs to handle all of these cases
properly.

This commit solves this by fixing estargz lib to handle entry paths with
cleaning them internally when it provides an unmarshaled view of an eStargz blob
through `estargz.Reader`. They are cleaned as relative to root, and estargz lib
APIs (e.g. `estargz.Reader.Lookup()`) also clean paths passed by the
client. This enables the client to read each TOCEntry.Name as a clean path and
enables to query TOCEntry without thinking about the differences of prefixes in
the marshalled TOC.

Note that paths are cleaned only in the `estargz.Reader` but not in
`estargz.Writer`. So TOC JSON created by estargz.Writer still contains
TOCEntry.Name identical to tar entry name (possibly prefixed with "/", "../",
"./" etc).

This commit also fixed tests to run for each prefix patterns in "", "./", "/"
and "../".

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>