estargz: avoid infinite loop of reference of entries
Currently, estargz lib makes an infinite reference loop of TOCEntry when the
input estargz contains the entry of the root directory itself. As a result, the
client walks infinite TOCEntry tree. This commit fixes this issue and adds a
test to detect it. This commit also adds a logic to prevent the infinite
recursive call of `cacheWithReader`.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>