Refactor keychain
We currently wrap `kubeconfig` keychain with ggcr and we use ggcr for getting
creds from `~/.docker/config.json`. Initially this was right because the
filesystem's auth logic was entiery based on ggcr.

But we refactored auth logic based on based on containerd's registry lib so it's
good time for us to refactor keychain without depending on ggcr as well.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>