Bump go to 1.15
When the snapshotter is compiled with especially the recent version of go, we
sometimes (or frequently) see that go-fuse prints `Write/Writev failed` log.
Here, I quote the log with few lines of debug logs around it.

```
2021/01/19 14:09:43 rx 76: FLUSH n2 {Fh 1}
2021/01/19 14:09:43 tx 76:     OK
...
2021/01/19 14:09:43 rx 77: INTERRUPT n0 {ix 76}
2021/01/19 14:09:43 tx 77:     11=resource temporarily unavailable
2021/01/19 14:09:43 writer: Write/Writev failed, err: 2=no such file or directory. opcode: INTERRUPT
```

This warning comes when the filesystem tries to write the response to INTERRUPT
(EAGAIN, in this case) to /dev/fuse [1].

When a filesystem operation of the client from userspace (`76: FLUSH`, in this
case) is interrupted by a signal, INTERRUPT request (`77: INTERRUPT` targetting
`76: FLUSH`, in this case) is sent to the FUSE server. In this case, there is a
race condition according to the kernel doc:

(Quoted from [2])

```
It is also possible that there's a race between processing the
original request and its INTERRUPT request.  There are two possibilities:

  1) The INTERRUPT request is processed before the original request is
     processed

  2) The INTERRUPT request is processed after the original request has
     been answered

If the filesystem cannot find the original request, it should wait for
some timeout and/or a number of new requests to arrive, after which it
should reply to the INTERRUPT request with an EAGAIN error.  In case
1) the INTERRUPT request will be requeued.  In case 2) the INTERRUPT
reply will be ignored.
```

As far as we can observe, we are currently seeing `Write/Writev failed` error on
case 2 where we writes EAGAIN to /dev/fuse. In the above log excerpt,
`77: INTERRUPT` is processed after `76: FLUSH` is returned with `OK`. Though
there is no document about errno in this case, FUSE seems to return ENOENT if
the interrupted request is not found on its side[3].

As a conclusion, this is an expected behaviour according to the kernel doc so we
can consider this phenomenon as harmless.

P.S. Similar phenomenon is observed on the internet[4], and they also conclude
this phenomenon as harmless.

[1] https://github.com/hanwen/go-fuse/blob/09a3c381714cf1011fb2d08885f29896cd496a0c/fuse/server_linux.go#L15
[2] https://www.kernel.org/doc/Documentation/filesystems/fuse.txt
[3] https://github.com/torvalds/linux/blob/219d54332a09e8d8741c1e1982f5eae56099de85/fs/fuse/dev.c#L1858-L1866
[4] https://bugzilla.redhat.com/show_bug.cgi?id=1763208#c17

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>