Preserve TOC item IDs when cloning memory metadata reader.
As the IDs are used for computing cache keys, it's important that they
do not change. Prior to this change, background fetching would clone the
reader and populate the cache with entries using possibly incorrect keys.

This patch changes the clone behavior to copy over the original ID
mappings.

Note that I also removed the locks around the ID maps as these maps are
never updated after the reader is created.

Fixes https://github.com/containerd/stargz-snapshotter/issues/842

Signed-off-by: Vadim Berezniker <vadim@berezniker.com>