CI: temporary disable containerd main branch integration test
containerd main branch changed plugin API that will be released in containerd v2
and that makes the plugin API not compatible to containerd v1 plugin APIs.
We'll migrate to the new API once containerd v2 is released.
As of now, just temporary disable the CI test against containerd main branch
until we migrate to v2 API.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>